+++ NEWS.md
@@ -1,3 +1,8 @@
+# SpiGesXML 0.0.7
+- Update default "latest" XSD schema for SpiGes 1.5
+- As the version number is not present anymore in the XSD schema, replace the
+  validation message in the console with the XSD schema URL.
+
 # SpiGesXML 0.0.6
 - add format validation message using spiges_get_df()
 - access IDs data with "Personenidentifikatoren" node